move class InverterIfc into a separate file
